Web page views moved to AlertaDengue/views.py (#141 fixes #142).
