+++ AlertaDengue/AlertaDengue/views.py
@@ -1,0 +1,58 @@
+from django.views.generic.base import TemplateView
+
+
+class MapaDengueView(TemplateView):
+    template_name = 'mapadengue.html'
+
+    def get_context_data(self, **kwargs):
+        context = super(MapaDengueView, self).get_context_data(**kwargs)
+        return context
+
+
+class MapaMosquitoView(TemplateView):
+    template_name = 'mapamosquito.html'
+
+    def get_context_data(self, **kwargs):
+        context = super(MapaMosquitoView, self).get_context_data(**kwargs)
+        return context
+
+
+class AboutPageView(TemplateView):
+    template_name = 'about.html'
+
+    def get_context_data(self, **kwargs):
+        context = super(AboutPageView, self).get_context_data(**kwargs)
+        # messages.info(
+        #   self.request,
+        #   'O site do projeto Alerta Dengue está em construção.')
+        return context
+
+
+class ContactPageView(TemplateView):
+    template_name = 'contact.html'
+
+    def get_context_data(self, **kwargs):
+        context = super(ContactPageView, self).get_context_data(**kwargs)
+        # messages.info(
+        #   self.request,
+        #   'O site do projeto Alerta Dengue está em construção.')
+        return context
+
+
+class JoininPageView(TemplateView):
+    template_name = 'joinin.html'
+
+    def get_context_data(self, **kwargs):
+        context = super(JoininPageView, self).get_context_data(**kwargs)
+        # messages.info(
+        #   self.request,
+        #   'O site do projeto Alerta Dengue está em construção.')
+        return context
+
+
+class PartnersPageView(TemplateView):
+    template_name = 'partners.html'
+
+    def get_context_data(self, **kwargs):
+        context = super(PartnersPageView, self).get_context_data(**kwargs)
+        return context
